+++ CHANGES.md
@@ -1,5 +1,6 @@
 ### Last Version
 
+  * Fix $path in schemas called by assert\all
   * Parameter extra for assert\dict accepts a schema dict
 
 ### 1.0.1 (2014-10-19)